projects/python-ocr/requirements.txt
easyocr
streamlit
opencv-python-headless
Pillow
numpy
